fix repoが逆
