+++ server/repository/gameRepository.ts
@@ -1,13 +1,16 @@
-import { userIdParser } from '$/service/idParsers';
-import type { Cell } from '@prisma/client';
+import { gameIdParser } from '$/service/idParsers';
+import type { Game } from '@prisma/client';
 import { z } from 'zod';
 
-const toCellModel = (prismaCell: Cell) => ({
-  where: z.object({ x: z.number().min(0), y: z.number().min(0) }).parse(prismaCell.where),
-  whoOpened: userIdParser.parse(prismaCell.whenOpened),
-  whenOpened: prismaCell.whenOpened?.getTime(),
+const toGameModel = (prismaGame: Game) => ({
+  id: gameIdParser.parse(prismaGame.id),
+  bombMap: z.array(z.array(z.union([z.literal(0), z.literal(1)]))).parse(prismaGame.bombMap),
+  userInputs: z
+    .array(z.array(z.union([z.literal(0), z.literal(1), z.literal(2)])))
+    .parse(prismaGame.bombMap),
 });
-export const cellRepository = {
+
+export const gameRepository = {
   create: () => {
     // /
   },
